// schemas/pet.ts
import { z } from 'zod'

export const createPetSchema = z.object({
  name: z.string(),
  nickName: z.string(),
  birthDate: z.date(),
  type: z.enum(['CAT', 'DOG']),
  gender: z.enum(['MALE', 'FEMALE', 'UNKNOWN', 'HERMAPHRODITE', 'ASEXUAL', 'NONE']),
  furType: z.enum(['SHORT', 'MEDIUM', 'LONG', 'HAIRLESS', 'NONE']),
  isReadyForAdoption: z.boolean().optional(),
  isFeatured: z.boolean().optional(),
  isAdopted: z.boolean().optional(),
  isPublished: z.boolean().optional(),
  isVaccinated: z.boolean().optional(),
  isSterilized: z.boolean().optional(),
  isTreatedForParasites: z.boolean().optional(),
  isLitterBoxTrained: z.boolean().optional(),
  isUsesScratchingPost: z.boolean().optional(),
  isSocialized: z.boolean().optional(),
  isFriendlyWithCats: z.boolean().optional(),
  isFriendlyWithDogs: z.boolean().optional(),
  isFriendlyWithOtherAnimals: z.boolean().optional(),
  status: z.enum(['ACTIVE', 'ARCHIVED', 'SUSPENDED']),
  profileId: z.string().cuid(),
  archivedReason: z.string(),
  suspendedReason: z.string(),
  colors: z.array(z.string()),
  bio: z.string().optional(),
  photos: z.array(z.object({})),
})

export const updatePetSchema = z.object({
  id: z.string().cuid(),
  name: z.string().min(1).optional(),
  nickName: z.string().min(1).optional(),
  birthDate: z.date().optional(),
  type: z.enum(['CAT', 'DOG']).optional(),
  gender: z.enum(['MALE', 'FEMALE', 'UNKNOWN', 'HERMAPHRODITE', 'ASEXUAL', 'NONE']),
  furType: z.enum(['SHORT', 'MEDIUM', 'LONG', 'HAIRLESS', 'NONE']),
  isReadyForAdoption: z.boolean().optional(),
  isFeatured: z.boolean().optional(),
  isAdopted: z.boolean().optional(),
  isPublished: z.boolean().optional(),
  isVaccinated: z.boolean().optional(),
  isSterilized: z.boolean().optional(),
  isTreatedForParasites: z.boolean().optional(),
  isLitterBoxTrained: z.boolean().optional(),
  isUsesScratchingPost: z.boolean().optional(),
  isSocialized: z.boolean().optional(),
  isFriendlyWithCats: z.boolean().optional(),
  isFriendlyWithDogs: z.boolean().optional(),
  isFriendlyWithOtherAnimals: z.boolean().optional(),
  status: z.enum(['ACTIVE', 'ARCHIVED', 'SUSPENDED']).optional(),
  profileId: z.string().cuid().optional(),
  archivedReason: z.string().optional(),
  suspendedReason: z.string().optional(),
  colors: z.array(z.string()).optional(),
  bio: z.string().optional().optional(),
  photos: z.array(z.object({})).optional(),
})

export type CreatePetData = z.infer<typeof createPetSchema>
export type UpdatePetData = z.infer<typeof updatePetSchema>